finished backend

+++ backend/api.go
@@ -5,20 +5,28 @@ import(
 	"net/http"
 	"encoding/json"
 	"context"
+    "time"
 
 	"github.com/Bryanthai/ordersystem/internal/database"
+    "github.com/Bryanthai/ordersystem/internal/auth"
 )
 
 const dbURL = "hahant:123456@tcp(localhost:3306)/fooddb?parseTime=true&tls=false"
+const authKey = "kp7Nbll8hKieGO2L1EQyOphkwJDVH0xD/tOp+DssAJ0Ll1t+jFnqdxK2BOmrAlMzQehX/2v4nt1xdDyXuU0CQQ=="
 
 // LOGIN
 func loginHandler(writer http.ResponseWriter, req *http.Request) {
     type LoginRequest struct {
         Username string `json:"username"`
+        Password string `json:"password"`
+        ExpiresInSeconds int `json:"expires_in_seconds"`
     }
     type LoginResponse struct {
-        Success bool   `json:"success"`
-        Message string `json:"message"`
+        Success  bool   `json:"success"`
+        Message  string `json:"message"`
+        Token    string `json:"token"`
+        ID       int32  `json:"id"`
+        Username string `json:"username"`
     }
 
     var loginReq LoginRequest
@@ -35,13 +43,29 @@ func loginHandler(writer http.ResponseWriter, req *http.Request) {
     defer db.Close()
 
     queries := database.New(db)
-    _, err = queries.GetAccount(context.Background(), loginReq.Username)
+    account, err := queries.GetAccount(context.Background(), loginReq.Username)
     if err != nil {
         http.Error(writer, "Invalid username", http.StatusUnauthorized)
         return
     }
+    err = auth.CheckPasswordHash(loginReq.Password, account.Password)
+    if err != nil {
+        http.Error(writer, "Invalid password", http.StatusUnauthorized)
+        return
+    }
 
-    resp := LoginResponse{Success: true, Message: "Login successful"}
+    if loginReq.ExpiresInSeconds <= 0 {
+        loginReq.ExpiresInSeconds = 3600
+    } else if loginReq.ExpiresInSeconds > 3600 {
+        loginReq.ExpiresInSeconds = 3600
+    }
+    token, err := auth.MakeJWT(account.ID, account.Username, time.Duration(loginReq.ExpiresInSeconds) * time.Second)
+    if err != nil {
+        http.Error(writer, "Error creating JWT", http.StatusInternalServerError)
+        return
+    }
+
+    resp := LoginResponse{Success: true, Message: "Login successful", Token: token, ID: account.ID, Username: account.Username}
     writer.Header().Set("Content-Type", "application/json")
     json.NewEncoder(writer).Encode(resp)
     return
@@ -51,6 +75,7 @@ func loginHandler(writer http.ResponseWriter, req *http.Request) {
 func registerHandler(writer http.ResponseWriter, req *http.Request) {
     type RegisterRequest struct {
         Username string `json:"username"`
+        Password string `json:"password"`
         Email    string `json:"email"`
         Address  string `json:"address"`
         Phone    int64	`json:"phone"`
@@ -63,6 +88,12 @@ func registerHandler(writer http.ResponseWriter, req *http.Request) {
     var regReq RegisterRequest
     if err := json.NewDecoder(req.Body).Decode(&regReq); err != nil {
         http.Error(writer, "Invalid request body", http.StatusBadRequest)
+        return
+    }
+
+    hashedPassword, err := auth.HashPassword(regReq.Password)
+    if err != nil {
+        http.Error(writer, "Error hashing password", http.StatusInternalServerError)
         return
     }
 
@@ -85,6 +116,7 @@ func registerHandler(writer http.ResponseWriter, req *http.Request) {
 
     err = queries.CreateAccount(context.Background(), database.CreateAccountParams{
         Username: 			regReq.Username,
+        Password: 			hashedPassword,
         Email:    			regReq.Email,
         Address:  			regReq.Address,
         UserPhoneNumber:	regReq.Phone,
@@ -104,6 +136,18 @@ func registerHandler(writer http.ResponseWriter, req *http.Request) {
 
 // ALTER ACCOUNT
 func alterAccountHandler(writer http.ResponseWriter, req *http.Request) {
+    token, err := auth.GetBearerToken(req.Header)
+    if err != nil {
+        http.Error(writer, "Invalid or missing token", http.StatusUnauthorized)
+        return
+    }
+
+    username, userID, err := auth.ValidateJWT(token, authKey)
+    if err != nil {
+        http.Error(writer, "Invalid token", http.StatusUnauthorized)
+        return
+    }
+
     type AlterAccountRequest struct {
         Username string `json:"username"`
         Email    string `json:"email"`
@@ -129,6 +173,13 @@ func alterAccountHandler(writer http.ResponseWriter, req *http.Request) {
     defer db.Close()
 
     queries := database.New(db)
+
+    account, err := queries.GetAccount(context.Background(), username)
+
+    if account.ID != userID || account.Username != username {
+        http.Error(writer, "Unauthorized", http.StatusUnauthorized)
+        return
+    }
     err = queries.AlterAccount(context.Background(), database.AlterAccountParams{
         Email:          alterReq.Email,
         Address:        alterReq.Address,
@@ -148,6 +199,18 @@ func alterAccountHandler(writer http.ResponseWriter, req *http.Request) {
 
 // CREATE NEW FOOD
 func createFoodHandler(writer http.ResponseWriter, req *http.Request) {
+    token, err := auth.GetBearerToken(req.Header)
+    if err != nil {
+        http.Error(writer, "Invalid or missing token", http.StatusUnauthorized)
+        return
+    }
+
+    username, userID, err := auth.ValidateJWT(token, authKey)
+    if err != nil {
+        http.Error(writer, "Invalid token", http.StatusUnauthorized)
+        return
+    }
+
     type CreateFoodRequest struct {
         FoodName    string  `json:"food_name"`
         FoodTag     string  `json:"food_tag"`
@@ -175,6 +238,12 @@ func createFoodHandler(writer http.ResponseWriter, req *http.Request) {
     defer db.Close()
 
     queries := database.New(db)
+
+    admin, err := queries.GetAccount(context.Background(), username)
+    if admin.ID != userID || admin.Username != username {
+        http.Error(writer, "Unauthorized", http.StatusUnauthorized)
+        return
+    }
 
 	var infoStruct sql.NullString
 	infoStruct.String = foodReq.Info
@@ -205,6 +274,18 @@ func createFoodHandler(writer http.ResponseWriter, req *http.Request) {
 
 // ALTER FOOD DATA
 func alterFoodHandler(writer http.ResponseWriter, req *http.Request) {
+    token, err := auth.GetBearerToken(req.Header)
+    if err != nil {
+        http.Error(writer, "Invalid or missing token", http.StatusUnauthorized)
+        return
+    }
+
+    username, userID, err := auth.ValidateJWT(token, authKey)
+    if err != nil {
+        http.Error(writer, "Invalid token", http.StatusUnauthorized)
+        return
+    }
+
     type AlterFoodRequest struct {
         FoodName    string  `json:"food_name"`
         FoodTag     string  `json:"food_tag"`
@@ -232,6 +313,11 @@ func alterFoodHandler(writer http.ResponseWriter, req *http.Request) {
     defer db.Close()
 
     queries := database.New(db)
+    admin, err := queries.GetAccount(context.Background(), username)
+    if admin.ID != userID || admin.Username != username {
+        http.Error(writer, "Unauthorized", http.StatusUnauthorized)
+        return
+    }
 
 	var infoStruct sql.NullString
 	infoStruct.String = foodReq.Info
@@ -261,6 +347,18 @@ func alterFoodHandler(writer http.ResponseWriter, req *http.Request) {
 
 // DELETE FOOD
 func deleteFoodHandler(writer http.ResponseWriter, req *http.Request) {
+    token, err := auth.GetBearerToken(req.Header)
+    if err != nil {
+        http.Error(writer, "Invalid or missing token", http.StatusUnauthorized)
+        return
+    }
+
+    username, userID, err := auth.ValidateJWT(token, authKey)
+    if err != nil {
+        http.Error(writer, "Invalid token", http.StatusUnauthorized)
+        return
+    }
+
     type DeleteFoodRequest struct {
         FoodName string `json:"food_name"`
     }
@@ -283,6 +381,11 @@ func deleteFoodHandler(writer http.ResponseWriter, req *http.Request) {
     defer db.Close()
 
     queries := database.New(db)
+    admin, err := queries.GetAccount(context.Background(), username)
+    if admin.ID != userID || admin.Username != username {
+        http.Error(writer, "Unauthorized", http.StatusUnauthorized)
+        return
+    }
     err = queries.DeleteFood(context.Background(), delReq.FoodName)
     if err != nil {
         http.Error(writer, "Failed to delete food item", http.StatusInternalServerError)
@@ -297,6 +400,19 @@ func deleteFoodHandler(writer http.ResponseWriter, req *http.Request) {
 
 // CREATE NEW ORDER
 func createOrderHandler(writer http.ResponseWriter, req *http.Request) {
+
+    token, err := auth.GetBearerToken(req.Header)
+    if err != nil {
+        http.Error(writer, "Invalid or missing token", http.StatusUnauthorized)
+        return
+    }
+
+    username, userID, err := auth.ValidateJWT(token, authKey)
+    if err != nil {
+        http.Error(writer, "Invalid token", http.StatusUnauthorized)
+        return
+    }
+
     type CreateOrderRequest struct {
         UserID    int32  `json:"user_id"`
         OrderInfo string `json:"order_info"`
@@ -321,6 +437,12 @@ func createOrderHandler(writer http.ResponseWriter, req *http.Request) {
     defer db.Close()
 
     queries := database.New(db)
+
+    account, err := queries.GetAccount(context.Background(), username)
+    if account.ID != userID || err != nil {
+        http.Error(writer, "Unauthorized", http.StatusUnauthorized)
+        return
+    }
     err = queries.CreateOrder(context.Background(), database.CreateOrderParams{
         UserID:    orderReq.UserID,
         OrderInfo: orderReq.OrderInfo,
@@ -339,6 +461,18 @@ func createOrderHandler(writer http.ResponseWriter, req *http.Request) {
 
 // GET ORDER STATUS
 func getOrderStatusHandler(writer http.ResponseWriter, req *http.Request) {
+    token, err := auth.GetBearerToken(req.Header)
+    if err != nil {
+        http.Error(writer, "Invalid or missing token", http.StatusUnauthorized)
+        return
+    }
+
+    username, userID, err := auth.ValidateJWT(token, authKey)
+    if err != nil {
+        http.Error(writer, "Invalid token", http.StatusUnauthorized)
+        return
+    }
+
     type GetOrderStatusRequest struct {
         UserID int32 `json:"user_id"`
     }
@@ -362,6 +496,12 @@ func getOrderStatusHandler(writer http.ResponseWriter, req *http.Request) {
     defer db.Close()
 
     queries := database.New(db)
+
+    account, err := queries.GetAccount(context.Background(), username)
+    if account.ID != userID || err != nil {
+        http.Error(writer, "Unauthorized", http.StatusUnauthorized)
+        return
+    }
     orders, err := queries.GetOrder(context.Background(), statusReq.UserID)
     if err != nil {
         http.Error(writer, "Failed to get orders", http.StatusInternalServerError)
@@ -379,6 +519,18 @@ func getOrderStatusHandler(writer http.ResponseWriter, req *http.Request) {
 
 // RATE ORDER
 func rateOrderHandler(writer http.ResponseWriter, req *http.Request) {
+    token, err := auth.GetBearerToken(req.Header)
+    if err != nil {
+        http.Error(writer, "Invalid or missing token", http.StatusUnauthorized)
+        return
+    }
+
+    username, userID, err := auth.ValidateJWT(token, authKey)
+    if err != nil {
+        http.Error(writer, "Invalid token", http.StatusUnauthorized)
+        return
+    }
+
     type RateOrderRequest struct {
         OrderID  int32  `json:"order_id"`
         Rating   int32  `json:"rating"`
@@ -403,6 +555,12 @@ func rateOrderHandler(writer http.ResponseWriter, req *http.Request) {
     defer db.Close()
 
     queries := database.New(db)
+
+    account, err := queries.GetAccount(context.Background(), username)
+    if account.ID != userID || err != nil {
+        http.Error(writer, "Unauthorized", http.StatusUnauthorized)
+        return
+    }
 
 	var feedbackStruct sql.NullString
 	feedbackStruct.String = rateReq.Feedback
@@ -429,6 +587,18 @@ func rateOrderHandler(writer http.ResponseWriter, req *http.Request) {
 
 // DELETE ORDER
 func deleteOrderHandler(writer http.ResponseWriter, req *http.Request) {
+    token, err := auth.GetBearerToken(req.Header)
+    if err != nil {
+        http.Error(writer, "Invalid or missing token", http.StatusUnauthorized)
+        return
+    }
+
+    username, userID, err := auth.ValidateJWT(token, authKey)
+    if err != nil {
+        http.Error(writer, "Invalid token", http.StatusUnauthorized)
+        return
+    }
+
     type DeleteOrderRequest struct {
         OrderID int32 `json:"order_id"`
     }
@@ -451,6 +621,12 @@ func deleteOrderHandler(writer http.ResponseWriter, req *http.Request) {
     defer db.Close()
 
     queries := database.New(db)
+    admin, err := queries.GetAccount(context.Background(), username)
+    if admin.ID != userID || admin.Username != username {
+        http.Error(writer, "Unauthorized", http.StatusUnauthorized)
+        return
+    }
+
     err = queries.DeleteOrder(context.Background(), delReq.OrderID)
     if err != nil {
         http.Error(writer, "Failed to delete order", http.StatusInternalServerError)
@@ -489,6 +665,56 @@ func getAllFoodHandler(writer http.ResponseWriter, req *http.Request) {
         Success: true,
         Foods:   foods,
         Message: "Food list retrieved successfully",
+    }
+    writer.Header().Set("Content-Type", "application/json")
+    json.NewEncoder(writer).Encode(resp)
+    return
+}
+
+func getAllOrdersHandler(writer http.ResponseWriter, req *http.Request) {
+    token, err := auth.GetBearerToken(req.Header)
+    if err != nil {
+        http.Error(writer, "Invalid or missing token", http.StatusUnauthorized)
+        return
+    }
+
+    username, userID, err := auth.ValidateJWT(token, authKey)
+    if err != nil {
+        http.Error(writer, "Invalid token", http.StatusUnauthorized)
+        return
+    }
+
+    type GetAllOrdersResponse struct {
+        Success bool              `json:"success"`
+        Orders  []database.Order  `json:"orders"`
+        Message string            `json:"message"`
+    }
+
+    db, err := sql.Open("mysql", dbURL)
+    if err != nil {
+        http.Error(writer, "Database error", http.StatusInternalServerError)
+        return
+    }
+    defer db.Close()
+
+    queries := database.New(db)
+
+    admin, err := queries.GetAccount(context.Background(), username)
+    if admin.ID != userID || err != nil {
+        http.Error(writer, "Unauthorized", http.StatusUnauthorized)
+        return
+    }
+    
+    orders, err := queries.GetAllOrders(context.Background())
+    if err != nil {
+        http.Error(writer, "Failed to get orders", http.StatusInternalServerError)
+        return
+    }
+
+    resp := GetAllOrdersResponse{
+        Success: true,
+        Orders:  orders,
+        Message: "Orders retrieved successfully",
     }
     writer.Header().Set("Content-Type", "application/json")
     json.NewEncoder(writer).Encode(resp)
